jinahub/encoders/text/LaserEncoder/gpu_requirements.txt
torch==1.9.0
laserembeddings[ja,zh]==1.1.1
jina-commons @ git+https://github.com/jina-ai/jina-commons.git#egg=jina-commons